requirements.txt
torch>=1.9.0
tqdm
albumentations>=1.0.3
opencv-python
Pillow>=8.3.1
numpy>=1.19.3
ttach
pandas
